This is the beginning for my first person shooter
In the beginning there was a blank slate. And that blank slate was a
derpy hello world program. This program will simple tell you hi, then
exiting back in to the dark abyss know as the system ram... Lol never
mind my cynical nature, this is the first commit and sync for my simple
first person shooter. This is a game project that I'll be working on for
a while to understand how first person shooter work. Its going to be a
very simple game were you are thrown into a randomly generated world.