SDL2 window and input implemented
I got a basic window framework for myself at the moment. It does basic
windowing, and it also allows inputs too.

The next thing to add is OpenGL 3.3. This is the base line version of
OpenGL I'll be using. In the near furture I'll also implement a OpenGL
2.0 render for the first person shooter game too so more computers can
run this simple FPS