Current Update 3 stuff
This is just a basic commit that I should have done a couple of days
ago. Or last night to be exact. This commit adds font, terrain, and a
new simple camera system to look around. The look around just press tab
key and it will go to camera look mode. The use the "wsad" keys to move
farward, backword, strafe left and right.  To move up or down press zx
keys.