Current Update 3 stuff the 2nd
I'm currently working on the lighting and material system. I've decided
to push back the animation format to a later milestone because I need to
seriously get done with these two graphical subsystesm and I need to
implement a geometry shader for the scene shader. I've also moved all of
the shader stuff to a centerlized "Renderer" class because I'm making
this game to be specific on how it renders stuff instead. The last two
big subsystems I need are lighting and the material system. I'm almost
done with lighting for now, then after that I'll focus on materials.