Update 4 first
This is what I've gotten done today. I've got the lighting system back
to were it was during the previous version. The code is actully shorter
now which is cool plus it simpliler than the microfacet model I was
using, however, subtract some of the features. I'm not ruling out a
microfacet lighting just putting a pause on it for now because I don't
really need to features of it just yet. I'm current working on the
material system. The material system will included 5 components I've
discussed the 5 components in the "Update 4 Lighting and Material." I'm
going to working on it tomorrow.