Update 4 Third
For now I'm done with the lighting and material system. I might do some
tweeking with it later, however, that will come after this update to the
game. The next thing I'll focus on for a week or so is to get some form
of shadow mapping in the game. Because I'm using 3 different models of
light I'm going to need to vary on the light types. I'll see later on.