Update 4 fourth
I've been working on the shadow portion of the project. It contains a
simple update for the first implementation of directional shadow
mapping. I've also added in several new objects including framebuffer,
drawsurface, and more. The render is coming a long smoothly compared to
my other renderers of the past. I think this one will defently be used
for this game. There is still much work that needs to be done.